Rename AuthenticationToken and add datetime comparison
